requirements.txt
asgiref==3.7.2
Django==4.2.6
django-ckeditor==6.7.0
django-js-asset==2.1.0
Pillow==10.1.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
